+++ CHANGES.rst
@@ -211,6 +211,9 @@ Smart Rollup node
 
 - Faster bootstrapping process. (MR :gl:`!8618`, MR :gl:`!8767`)
 
+- Added a new metrics ``head_inbox_process_time`` to report the time the rollup
+  node spent to process a new Layer 1 head. (MR :gl:`!8971`)
+
 Smart Rollup client
 -------------------